backend/requirements.txt
fastapi
uvicorn
numpy
python-multipart
aiofiles
python-dotenv
cloudinary
motor
pydantic
transformers
tokenizers
torch
Pillow
librosa
ffmpeg-python
moviepy
requests
typing-extensions
opencv-python
scikit-image
google-generativeai
python-magic; sys_platform != "win32"       # On Linux and MacOS install python-magic
python-magic-bin; sys_platform == "win32"   # On Windows install python-magic-bin
time